+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * each-else
 * else-if
 * refactoring $onMount
+* optional deep checking for passed props: prop={value} prop|deep={value}
 
 ## 0.6.x